Portfolio Demonstration
The following images display the work that I have put in for the School Simplified and their child organisation Programming Simplified as a volunteer web designer on their web development team, which involved liaising with front-end engineers and stakeholders to display the NPO's brand perfectly. Made use of Figma, Canva and developed in JavaScript, HTML, CSS, Chakra-UI and React.